Add documentation to allow #![deny(missing_docs)].
Add documentation to all functions, structs, and types to satisfy
the linter when #[deny(missing_docs)] is used. Module-level
documentation is still not included for the internal modules since
it's really redundant at the moment, so #[deny(missing_docs)] is
not currently enabled by default. The documentation for many
functions are quite terse and could use additional information to
make them accessible for end-users of the library.