Add benchmarking tests for Parser.
Add a benchmarking test for the parser to detect performance
regressions during development. The Criterion benchmark crate is
used to perform the benchmarks and provides a number of useful
metrics for evaluating changes between runs.

To run the benchmarks, run the following command:

  cargo bench -- --test [--nocapture]

This has been added to the .travis.yml file so that performance
benchmarks are run as part of the standard suite of continuous
integration tests.

In order to preserve benchmarking comparisons between releases, the
.criterion directory is checked into git, but is also in .gitignore.
This allows stable performance comparisons between releases and
during CI runs, rather than being overwritten with each checkin.

A new version of the .criterion directory should be checked into git
manually at each release to capture the new performance baseline
for future development.